Made vim highlighting work mostly how I want now.
I made `cpp.cplusplus` a hybrid filetype which tricks the vim
highlighting into believing that we're really in raw C++ for the
purposes of the core vim highlighting rules.  My `cplusplus` type
adds more bits to the core stuff, but mostly exists to let
doxygen slip in the way that I want it to.